Use the iterator verifier and factor out common code for dotproduct and weighted set.
